Fixed issue where EB Dev environment was unable to use environment variables
Description of bug/oversight:

The development server version of the app (localhost) has a different behavior when it comes to using environment variables. the .env file is able to be reached from both client and server.

However the production build of the app (EB dev environment) separates into two environments: static user interface (client) and runtime (server)

This means getServerSideProps is required to fetch process.env from the server in the client pages.

An adjustment to how public and private encryption keys was also necessary for it to be stored and retrieved from eb env variable storage in the proper format. Currently its stored in base64 and decoded into utf-8 before use.